feat: Reservation Create, Delete, Get API 로직 추가

+++ src/main/java/finalmission/service/ReservationTimeService.java
@@ -3,7 +3,9 @@ package finalmission.service;
 import finalmission.domain.ReservationTime;
 import finalmission.dto.request.ReservationTimeCreateRequest;
 import finalmission.dto.response.ReservationTimeResponse;
+import finalmission.exception.InvalidRequestException;
 import finalmission.exception.NotFoundException;
+import finalmission.repository.ReservationRepository;
 import finalmission.repository.ReservationTimeRepository;
 import java.util.List;
 import org.springframework.stereotype.Service;
@@ -11,9 +13,12 @@ import org.springframework.stereotype.Service;
 @Service
 public class ReservationTimeService {
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeService(final ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(final ReservationTimeRepository reservationTimeRepository,
+                                  final ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public ReservationTimeResponse create(ReservationTimeCreateRequest request) {
@@ -21,10 +26,15 @@ public class ReservationTimeService {
         return ReservationTimeResponse.from(savedReservationTime);
     }
 
-    public void removeReservation(final Long id) {
+    public void removeReservationTime(final Long id) {
         if (!reservationTimeRepository.existsById(id)) {
             throw new NotFoundException();
         }
+
+        if (reservationRepository.findByTimeId(id)) {
+            throw new InvalidRequestException("예약이 존재하고 있어 삭제할 수 없습니다");
+        }
+
         reservationTimeRepository.deleteById(id);
     }
 
@@ -33,4 +43,10 @@ public class ReservationTimeService {
                 .map(ReservationTimeResponse::from)
                 .toList();
     }
+
+    public ReservationTime find(Long id) {
+        return reservationTimeRepository.findById(id)
+                .orElseThrow(NotFoundException::new);
+    }
+
 }
